+++ db/migrate/20170509183608_add_release_date_to_story.rb
@@ -1,0 +1,5 @@
+class AddReleaseDateToStory < ActiveRecord::Migration
+  def change
+    add_column :stories, :release_date, :date
+  end
+end